feat: bigint로 변경

+++ src/main/java/com/chan/stock_portfolio_backtest_api/stock/domain/StockPrice.java
@@ -38,7 +38,7 @@ public class StockPrice {
 	private Float lowPrice;
 	@Column(nullable = false)
 	private Float highPrice;
-	private Integer tradeQuantity;
+	private Long tradeQuantity;
 	private Long tradeAmount;
 	private Long issuedCount;
 	@Column(nullable = false)
